Move createPlayerCollector to its own file
This'll be on its own file in `interactionHandlers.js`. I foresee
myself making more 'general' functions here :). Potentially renaming
this file as well 💀. This function becomes createCollector() and
has an additional parameter which is a function that's executed
when a collector collects something.
